requirements.txt
pyserial>=3.5
flask>=2.3.0
pyarrow>=12.0.0
